refactor: type the serialized command log and relocate CommandEntry to the tap types
