+++ packages/app/src/tap/commands/commands.cy.ts
@@ -38,7 +38,6 @@ describe('tap/commands/commands', () => {
   // the runner seam rather than replace it.
   const stubRunner = (runner: unknown) => cy.stub(tapManagerDataSource, 'getRunner').returns(runner)
 
-  // Mirror the driver: getTestState resolves one test by id, or undefined.
   const getTestStateFrom = (state: Record<string, unknown>) => cy.stub().callsFake((id: string) => state[id])
 
   it('fails with NO_RUN when no spec has mounted a runner yet', async () => {
