+++ packages/driver/src/cypress/runner.ts
@@ -1955,8 +1955,6 @@ export default {
 
         const test = serializeTest(testRunnable)
 
-        // `_titlePath` is only stamped on the normalized runnable copy;
-        // `getTestById` resolves the raw runnable, so read it off the live one.
         test._titlePath = testRunnable.titlePath()
 
         test.prevAttempts = _.map(testRunnable.prevAttempts, serializeTest)
